+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.4.0] - Unreleased
+### Added
+- Support for negative filters [#178].
+  Example: `{{ value |> !empty }}`
+
 ## [2.3.1] - 2026-02-21
 ### Fixed
 - Types for Node [#174]
@@ -122,7 +127,9 @@ Vento 2.0 is now dependency-free and compatible with browsers without a build st
 [#166]: https://github.com/ventojs/vento/issues/166
 [#167]: https://github.com/ventojs/vento/issues/167
 [#174]: https://github.com/ventojs/vento/issues/174
+[#178]: https://github.com/ventojs/vento/issues/178
 
+[2.4.0]: https://github.com/ventojs/vento/compare/v2.3.1...HEAD
 [2.3.1]: https://github.com/ventojs/vento/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/ventojs/vento/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/ventojs/vento/compare/v2.1.1...v2.2.0